ss-main: fixed the files
