+++ global_chem/global_chem/food/fruits/mango/mango_phenolic_acids.py
@@ -1,3 +1,8 @@
+#!/usr/bin/env python3
+#
+# GlobalChem - Mango Phenolic Acids
+# ----------------------------------
+
 class MangoPhenolicAcids(object):
 
     def __init__(self):
